backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
mido==1.3.0
python-rtmidi==1.5.8